chore: add query for classification
