+++ src/app/[...slug]/page.tsx
@@ -77,6 +77,7 @@ export default async function Page({ params, searchParams }: PageParams) {
     let appState : AppState = {
         requestContext: {
             pageNode: layout,
+            searchParams: searchParams,
             detailItem: layout.DetailItem,
             culture: layout.Culture,
             isEdit: searchParams["sfaction"] === "edit",
