+++ src/components/call-to-action/call-to-action.tsx
@@ -3,10 +3,9 @@ import { StyleGenerator } from "../styling/style-generator.service";
 import { OffsetStyle } from "../styling/offset-style";
 import { Alignment } from "../styling/alignment";
 import { ButtonType } from "../styling/button-types";
-import { generateAnchorAttrsFromLink, htmlAttributes } from "sitefinity-react-framework/widgets/attributes";
+import { generateAnchorAttrsFromLink, getCustomAttributes, htmlAttributes } from "sitefinity-react-framework/widgets/attributes";
 import { classNames } from "sitefinity-react-framework/utils/classNames";
 import { WidgetContext } from "sitefinity-react-framework/widgets/widget-context";
-import { CustomAttribute } from "../styling/custom-attribute";
 import { LinkModel } from "sitefinity-react-framework/interfaces/LinkModel";
 
 export type CTAPart = "Wrapper" | "Primary" | "Secondary";
@@ -81,15 +80,4 @@ export class CallToActionEntity {
     };
     CssClass?: string;
     Margins?: OffsetStyle;
-}
-
-const getCustomAttributes = (attributes: any, part: CTAPart) => {
-    if (!attributes || !attributes[part]){
-        return {};
-    }
-
-    return attributes[part].reduce(
-        (current: object, customAttribute: CustomAttribute) => {
-            return ({ ...current, [customAttribute.Key]: customAttribute.Value})
-        }, {});
 }
