--- src/components/styling/custom-attribute.ts
@@ -1,4 +1,0 @@
-export type CustomAttribute = {
-    Key: string;
-    Value: string;
-};
